[ADD] module to work on annotations over a genome.
This stuff is taken from my other project guizmin. Basically Genome contains two modules. The first is Location, which is similar to Range, apart that it contains a chromosome name in addition to the interval bounds. The second is Annotation and is really similar to RSet. There are a couple of important distinctions though:
- it contains [Location]s instead of [Range]s
- each [Location] is associated to a value (the data structure is polymorphic with respect to these location-attached values)
- it is based on arrays instead of lists, which provides search in O(log(n)) where n is the number of locations contained in the annotation. I used it for working with reads in sequencing data, and it works ok (but it could easily be improved i think)