Zip.unzip: implement "Gzip concatenation"
In order to be compliant with the RFC 1952 (and hence to be able to
parse BAM files), one has to implement the ability to decode
concatenated Gzip streams (the uncompressed stream must be the
concatenation of the uncompressed initial files).

The test uses many different buffer sizes because many
buffer-arithmetic bugs were discovered during development.