Bam: draft a first version of the BAM parser
The parser is still fragile and low-level but it seems to have
successfully parsed one big BAM file (2 640 496 alignment records).
The global reference Biocaml_bam.debug will be removed when we have
more confidence on the code, right now it defaults to `true' and
implies printing a lot of debug information on `stderr'.
There are no automatic tests so far.