Jaspar: don't eval regexp at module load time
This was done to work around Issue #56, but is perhaps a good idea
anyway. However, #56 is not resolved, just avoided. Calling this
function will still cause the error at runtime. I'm not sure how to
fix this, and perhaps the real blame lies in the version of PCRE on my
system.