Suppress the dependency on pcre (fix #56)
The last (only?) module using pcre was Jaspar. When trying to remove
the use of pcre in this module, I noticed the layout of the database
had recently been significantly changed. So I rewrote the whole
module, taking care of not using pcre this time. I also removed all
occurrences of pcre in the various configuration scripts.

Last point: I wrote a test for the parser, but to run it with other
unit tests I need a chunk of jaspar, and for licensing issues I
suppose it is not allowed to include a part of jaspar in the biocaml
distribution. So the unit test is not included in this patch.