Sam: completely new implementation
New implementation based on Or_error and Future. Type definitions, a
major component of this module, are much more precise now. Structured
both the mli and ml in clear separate sections for type defs, parsing
functions, and IO.

Added dependency on Re library, which is being used in the new
implementation.

Previous module not deleted, but just renamed to Sam_deprecated. This
is now being used in dependent modules to avoid having to change too
much in one commit.