opam: add version constraint on camlzip
We require the `zip` findlib package, which is not provided by camlzip
1.04 (as mentioned in #144). We could instead require `camlzip`, which
would make us compatible with all versions of camlzip. However, it
seems 1.05 is deprecating the `camlzip` findlib package name in favor
of `zip` (because camlzip is defined as an alias to zip, the library
file is zip.cma). Thus, might as well go with that.