stop compiling Tags and most apps
We are making broad changes and it was difficult to do so since Tags
depends on every module. Removing it and all the apps that depend on it
will simplify development at this time.

All affected files moved into a tmp/ directory, so we can still
remember the code that needs to be re-incorporated.