base/Fasta: changed API for Parser submodule
`Parser.step` now deals with `string` inputs instead of `items0` and
uses `Parser0.step`. This is more appropriate for external uses, and
the API makes more sense, notably for error handling. In addition the
implementation is also simpler, because the high-level parser can
benefit from properties guaranteed by the low-level parser: the latter
is guaranteed to produce a valid sequence of items (no consecutive
description lines for instance)