removed dependency on future
Removing dependency on future was achieved by copying (most of) the
library's contents into biocaml. Some parts of the original future
signature (Sys module and Unix stat type) were removed because they
are not implemented in core_kernel and also not used in biocaml.

Also removed the (hopefully) last remnants of core in biocaml.unix,
and biocaml.test.